make lcdknob work!
Finally, we're having some real action here. "Hello!" being printed to
the screen. For some reason though, there's a space being inserted in
the middle of the word. No idea why yet.

A big advantage we have here compared to Arduino's "LiquidCrystal"
library is that we properly use the BF flag so that we don't have to
wait for the screen *maximal* command process time. We can wait for
precisely the time it takes for the command to be processed.